Revert "Fixed Boiler Plate, Implements Player Object"
