+++ classes/jsonHandler.js
@@ -1,0 +1,17 @@
+module.exports = {
+    createBlackCard: function (user, io) {
+
+    },
+
+    createWhiteCard: function (user, io) {
+
+    },
+    createCurrentDisplay: function (user, io) {
+
+    },
+    createPlayer: function (user, io) {
+
+    }
+}
+
+// Each function takes the parameters required to create the respective objects and return the objects created.
